+++ src/api/create-dashboard.js
@@ -3,7 +3,7 @@ var cdb = require('cartodb.js');
 var Dashboard = require('./dashboard');
 var DashboardView = require('../dashboard-view');
 var WidgetsCollection = require('../widgets/widgets-collection');
-var WidgetsService = require('./widgets-service');
+var WidgetsService = require('../widgets-service');
 
 /**
  * Translates a vizJSON v3 datastructure into a working dashboard which will be rendered in given selector.
